+++ src/accounts/Account.java
@@ -2,7 +2,4 @@ package accounts;
 
 public class Account {
     
-    public Account () {
-
-    }
 }