+++ src/users/Admin.java
@@ -1,28 +1,13 @@
 package users;
 
 public class Admin extends User {
+    private String name;
+    private String email;
+    private String password;
+    private int id;
+    
     public Admin(String name, String email, String password, int id) {
         super(name, email, password, id);
-    }
-    @Override
-    public String getName() {
-        return super.getName();
-    }
-    @Override
-    public String getEmail() {
-        return super.getEmail();
-    }
-    @Override
-    public String getPassword() {
-        return super.getPassword();
-    }
-    @Override
-    public int getId() {
-        return super.getId();
-    }
-    @Override
-    public void changePassword(String newPassword) {
-        super.changePassword(newPassword);
     }
 }