+++ src/util/FileIO.java
@@ -1,9 +1,12 @@
 package util;
 import java.io.File;
 import java.io.FileWriter;
-import users.*;
+import java.io.IOException;
 
 public class FileIO {
+    public static String getDatabase() {
+        return "/Users/xpeezy/Desktop/CS/VS code proj/OOP practice/bank/customer_database.txt";
+    }
     public static void createFile(String filename) {
         try {
             File file = new File(filename);
@@ -11,23 +14,14 @@ public class FileIO {
                 System.out.println("\nFile " + filename + " created at " + file.getAbsolutePath());
             else
                 System.out.println("\nWriting to " + filename + "...");
-        } catch (Exception e) {
+        } catch (IOException e) {
             System.out.println("Error creating file: " + e.getMessage());
         }
     }
-    public static void readFile(String filename) {
-        try (java.util.Scanner scanner = new java.util.Scanner(new File(filename))) {
-            while (scanner.hasNextLine()) {
-                System.out.println(scanner.nextLine());
-            }
-        } catch (Exception e) {
-            System.out.println("Error reading file: " + e.getMessage());
-        }
-    }
     public static void writeFile(String filename, String content) {
-        try (FileWriter writer = new FileWriter(filename)) {
+        try (FileWriter writer = new FileWriter(filename, true)) {
             writer.write(content);
-        } catch (Exception e) {
+        } catch (IOException e) {
             System.out.println("Error writing to file: " + e.getMessage());
         }
     }
@@ -39,19 +33,6 @@ public class FileIO {
             System.out.println("Failed to delete file.");
         }
     }
-    public static boolean searchString(String filename, String searchTerm) {
-        try (java.util.Scanner scanner = new java.util.Scanner(new File(filename))) {
-            while (scanner.hasNextLine()) {
-                String line = scanner.nextLine();
-                if (line.contains(searchTerm)) {
-                    return true;
-                }
-            }
-        } catch (Exception e) {
-            System.out.println("Error searching file: " + e.getMessage());
-        }
-        return false;
-    }
     public static boolean searchInt(String filename, int searchNumber) {
         try (java.util.Scanner scanner = new java.util.Scanner(new File(filename))) {
             while (scanner.hasNextLine()) {
@@ -59,17 +40,17 @@ public class FileIO {
                 if (line.contains(String.valueOf(searchNumber)))
                     return true;
             }
-        } catch (Exception e) {
+        } catch (IOException e) {
             System.out.println("Error searching file: " + e.getMessage());
         }
         return false;
     }
-    public static String getCustomerData(Customer currentUser) {
+    public static String formatCustomerData(String name, String password, int id) {
             return    "-------------------------------------\n"
-            + Console.getCurrentTimeAndDate() + "\n"
+            + Utils.getCurrentTimeAndDate() + "\n"
             + "-------------------------------------\n"
-            + "Name: " + currentUser.getName() + "\n"
-            + "Password: " + currentUser.getPassword() + "\n"
-            + "ID #: " + currentUser.getId() + "\n\n";
+            + "Name: " + name + "\n"
+            + "Password: " + password + "\n"
+            + "ID #: " + id + "\n";
     }
 }