+++ src/views/View.java
@@ -4,23 +4,13 @@ import java.util.Scanner;
 
 public abstract class View {
 
-    protected Scanner scan;
+    protected Scanner viewScanner;
 
-    protected View(Scanner scan) {
-        this.scan = scan;
+    protected View(Scanner mainScanner) {
+        viewScanner = mainScanner;
     }
-    public void show() {};
-    public void show(View parent) {};
-    protected void returnToWelcomeView() {
-        scan.nextLine(); // Consume the newline left by nextDouble()
-        while (true) {
-            if (scan.nextLine().isEmpty()) {
-                new WelcomeView(scan).show();
-                break;
-            } 
-            else {
-                System.out.println("Invalid input, press ENTER.");
-            }
-        }
+    public abstract View show();
+    public Scanner getScanner() {
+        return viewScanner;
     }
 }